fix(data-drains): extract sleepUntilAborted, honor abort across all destinations
